+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [unreleased]
 
+- [#2251](https://github.com/teambit/bit/issues/2251) support css url functions without explicitly import
+
 ## [2.1.5] - 2020-01-12
 
 ### Bug fixes